back to $fa slices, not helped much
My findings:
- with 2014.QX speed can be improved much by using list comprehensions.
But I think a library should not expect beta versions.
- to create a correct profile it is needed that the plane step angles
are irregular because thread flank corners angles do not match plane
step angles. But this confuses linear_extrude() and only renders with
high $fn are usable. ==> Sadly, the concept of extruding twisted plane
is probably impossible to display nicely in this version of OpenScad. If
someone finds a fix for this. You are welcome